Software changes to improve data reading rate, changes to MATLAB
By default, data from device is retrieved via AXI and then saved
to a binary file.  This binary file ("SavedData.bin") is loaded
by Python before being sent over TCP/IP to the client computer.
This improves the data transfer rate by an enormous factor. Max
data sizes that can be transferred are now limited by the
connection timeout and not by anything else.

Changed the MATLAB function names for reading data to getRAM and
getFIFO.  Can now also retrieve the current ADC data (without
filtering) using the method readADC.

Had some trouble with calibration of the device, and I'm not
sure why.  I think the issue was that the board I was using
is the board with some kind of electrical fault.

Also changed the names of the C functions to fetchRAM.c and
fetchFIFO.c.