Added lock-in detection module, changed fast filtering
Fast filtering has been changed slightly.  Previously, the filter
started averaging when the trigger was received.  Now it averages
all the time, which may mean that the timing information about when
a signal occurs will be undetermined up to the sample period.  I
made this change because it was easier to use the new SaveADCData
module, and also I will at some point move to a CIC filter to get
better rejection of high frequencies, and this filter will always
be active.

Added a lock-in detection module.  This was taken mostly from the
laser servo project, and it is still untested.  MATLAB and C code
has been added to make it work.