Fixed many bugs with lock-in detection, updated README
Had a lot of issues with lock-in detection and also the changes
I made to how data is saved to the memory.  Had the memory read
logic connected to the wrong signals, which caused the device to
hang.  Demodulated signals were not divided down enough.  Other
annoying things like that.

Lock-in has been tested by connecting the DAC output to the input
and looking at the signals.  Seems to work.