Fixed DDS reset issue, device now works
Problem with reset was that the Streaming phase DDS IP, which does
the mixing, had the wrong reset signal wired up and thus was always
in a permanent state of reset.

Changed default FPGA registers so that the external trigger is
always enabled and the default edge is rising.  Default number
of samples is now 16380, and the default delay is now 100 clock
edges.