Added digital pattern generator for gating signal acquisition
Digital pattern generator with 256 addresses and 8 bits of data
has been added to allow for gating of signal acquisition.  Software
for writing gating signals to controller has been added.  Memory
resets added, and these should be called prior to running a gated
acquisition because the number of points acquired may not be the
total number requested, and the saving module will not complete.

Design has been tested and it seems to work